refactor: centralize get_historical_vix
